requirements.txt
pyarrow==12.0.0
pandas~=1.5.3
lightgbm==3.3.5
pyyaml==6.0
scikit-learn~=1.2.2
optuna~=3.1.1
joblib~=1.1.0
numpy~=1.24.3
streamlit==1.22.0
matplotlib==3.7.1
plotly==5.14.1
langid==1.1.6